feat: extract UseTendrilProcessView hook to deduplicate process view creation
Replace inline TendrilProcessView + CreatePlanDialogLauncher construction
in WallpaperApp, DraftsApp, and ReviewApp with a shared UseTendrilProcessView
hook that encapsulates status subscription and widget building.
